📦 Refactor code + 📝 Some documentation
