+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "firebase": "^6.4.2",
+    "lodash": "^4.17.15",
     "memoize-one": "^5.1.1",
     "prop-types": "^15.7.2",
     "react": "^16.9.0",
